[FIX] 잘못된 엔티티 ID 타입 수정

+++ src/main/java/org/websoso/WSSServer/domain/AttractivePoint.java
@@ -18,7 +18,7 @@ public class AttractivePoint {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(nullable = false)
-    private Long attractivePointId;
+    private Byte attractivePointId;
 
     @Column(columnDefinition = "varchar(12)", nullable = false)
     private String attractivePointName;
